pyproject.toml
[project]
name = "adk-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.10.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.10.0,<4",
    "google-genai>=1.29.0",
]
